refactor: repo browser の branch orchestration を分離する
